DDP Buffering Tests for Meteor 1.3.3
This is an adaptation of @tmeasday's 74230beba8777cb9104c4180525e2a3fe09772c4 test that he created for meteor/meteor#5680.  Due to occasional failures, it now uses sinon's `lolex` npm library to allow us to control the setTimeout/setInterval within the test itself, providing for tick-accurate testing. Also nifty because it allows the test to finish in less time than it actually takes.